# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Open-Voice-Journal API
  version: 0.3.2
  description: API contract for recording ingestion, uploads, provider configuration, and summaries.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register a new user account (when open signup is enabled)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                displayName:
                  type: string
      responses:
        '201':
          description: User created
        '403':
          description: signup_disabled
...